CP-54351 move SR from /opt
