+++ Makefile
@@ -54,6 +54,7 @@ SM_LIBS += nfs
 SM_LIBS += pluginutil
 SM_LIBS += refcounter
 SM_LIBS += resetvdis
+SM_LIBS += SR
 SM_LIBS += SRCommand
 SM_LIBS += srmetadata
 SM_LIBS += sysdevice
