+++ drivers/DummySR.py
@@ -17,7 +17,7 @@
 #
 # DummySR: an example dummy SR for the SDK
 
-import SR
+from sm import SR
 from sm import VDI
 from sm import SRCommand
 from sm.core import util
