+++ drivers/EXTSR.py
@@ -17,8 +17,8 @@
 #
 # EXTSR: Based on local-file storage repository, mounts ext3 partition
 
-import SR
-from SR import deviceCheck
+from sm import SR
+from sm.SR import deviceCheck
 from sm import SRCommand
 import FileSR
 from sm.core import util
