+++ drivers/NFSSR.py
@@ -19,7 +19,7 @@
 
 import socket
 
-import SR
+from sm import SR
 from sm import SRCommand
 import FileSR
 from sm.core import util
