+++ drivers/SMBSR.py
@@ -17,7 +17,7 @@
 #
 # SMBSR: SMB filesystem based storage repository
 
-import SR
+from sm import SR
 from sm import SRCommand
 import FileSR
 from sm.core import util
